获取空间中的所有设备

+++ src/main/java/edu/fudan/se/sctap_lowcode_tool/service/impl/SpaceServiceImpl.java
@@ -4,11 +4,12 @@ import edu.fudan.se.sctap_lowcode_tool.model.DeviceInfo;
 import edu.fudan.se.sctap_lowcode_tool.model.SpaceInfo;
 import edu.fudan.se.sctap_lowcode_tool.repository.SpaceRepository;
 import edu.fudan.se.sctap_lowcode_tool.service.SpaceService;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.Optional;
 import java.util.Set;
-
 
 @Service
 public class SpaceServiceImpl implements SpaceService {
@@ -17,31 +18,40 @@ public class SpaceServiceImpl implements SpaceService {
     private SpaceRepository spaceRepository;
 
     @Override
-    public SpaceInfo saveSpaceInfo(SpaceInfo spaceInfo) {
+    public SpaceInfo saveOrUpdateSpace(SpaceInfo spaceInfo) {
         return spaceRepository.save(spaceInfo);
     }
 
     @Override
-    public SpaceInfo getSpaceInfoById(int spaceId) {
-        return spaceRepository.findById(spaceId).orElse(null);
+    public boolean deleteSpace(int spaceId) {
+        if (spaceRepository.existsById(spaceId)) {
+            spaceRepository.deleteById(spaceId);
+            return true;
+        }
+        return false;
     }
 
     @Override
-    public SpaceInfo updateSpaceInfo(SpaceInfo spaceInfo) {
-        return spaceRepository.save(spaceInfo);
+    public Optional<SpaceInfo> findSpaceById(int spaceId) {
+        return spaceRepository.findById(spaceId);
     }
 
     @Override
-    public void deleteSpaceInfoById(int spaceId) {
-        spaceRepository.deleteById(spaceId);
+    public Set<DeviceInfo> getAllSpaceDevices(int spaceId) {
+        return spaceRepository.findById(spaceId)
+                .map(SpaceInfo::getSpaceDevices)
+                .orElseThrow(() -> new EntityNotFoundException("No any devices in space with id: " + spaceId));
     }
 
-//    @Override
-//    public Set<DeviceInfo> getAllSpaceDevices(int spaceId) {
-//        SpaceInfo spaceInfo = spaceRepository.findById(spaceId).orElse(null);
-//        if (spaceInfo != null) {
-//            return spaceInfo.getDevices();
-//        }
-//        return null;
-//    }
+    @Override
+    public int getSpaceIdByName(String spaceName) {
+        return spaceRepository.findIdBySpaceName(spaceName)
+                .orElseThrow(() -> new EntityNotFoundException("Space not found with name: " + spaceName));
+    }
+
+
+    @Override
+    public Iterable<SpaceInfo> findAllSpaces() {
+        return spaceRepository.findAll();
+    }
 }
